requirements.txt
peewee
pendulum
mysqlclient
redis
flask-redis
flask
requests
pycryptodome
PyJWT